WebUI 2 uses a very strange code style but only because it's designed for speed of coding and not end production. Don't worry, it will be completely different before you know it.
Usage:
Search for the name of a key, for example blog. The enter button clicks the button as you would expect.
Click on friends to see a list of friends. Their names are currently Anonymous, which needs to be fixed.
Click on collections to see a list of all collections.
Click on Add content to create a new piece of content. The subject is the key. It is currently the only way to search for a post. This will go away in the future to support standardized naming conventions and full text search.
Click on Add friend to add a friend using their passport.
Click on an x to remove a friend or delete content.
Click on the randimage in the upper right corner to get your own info. Click on the box to put it into the feed.

TODO:
Names don't work at the moment, this is first to be fixed.
Full-text search.
Download all collection keys at the very start. Images are costly, so delay downloading of them for a second.
Cannot change your own name.
Cannot change the structure of data posted to the content.
Test for XSS, CSRF (the api itself is vulnerable to this), and so forth.
Upgrade to jQuery 2.x if you want to (the only difference is that jQuery 2.x doesn't support IE6).
Improve RandImages to look more like avatars.
Improve the UX.
Improve the asthetic.
Allow the user to easily turn off the easter egg fire logo. Currently possibly only by adding #no to the url.